+++ nautobot_data_validation_engine/__init__.py
@@ -11,7 +11,6 @@ __version__ = metadata.version(__name__)
 import logging
 from django.db.utils import ProgrammingError
 from nautobot.extras.plugins import NautobotAppConfig, register_template_extensions
-from nautobot.extras.plugins.utils import import_object
 
 logger = logging.getLogger(__name__)
 
@@ -30,31 +29,27 @@ class NautobotDataValidationEngineConfig(NautobotAppConfig):
     max_version = "1.9999"
     default_settings = {}
     caching_config = {}
-    audit_rulesets = "audit_rulesets.audit_rulesets"
 
     def ready(self):
         super().ready()
         from nautobot_data_validation_engine.template_content import tab_factory  # pylint: disable=C0415
         from django.contrib.contenttypes.models import ContentType  # pylint: disable=C0415
 
-        audit_rulesets = import_object(f"{self.__module__}.{self.audit_rulesets}")
-        if audit_rulesets is not None:
-            self.features["audit_rulesets"] = sorted(set(audit_ruleset.model for audit_ruleset in audit_rulesets))
-            template_content = []
-            labels = []
-            try:
-                for content_type in ContentType.objects.all():
-                    label = f"{content_type.app_label}.{content_type.model}"
-                    labels.append(label)
-                    template_content.append(tab_factory(label))
-                register_template_extensions(template_content)
-                self.features["template_extensions"] = sorted(set(labels))
-            except ProgrammingError:
-                logger.warning(
-                    "Creating template content for validation engine failed because "
-                    "the ContentType table was not available or populated. This is normal "
-                    "during the execution of the migration command for the first time."
-                )
+        template_content = []
+        labels = []
+        try:
+            for content_type in ContentType.objects.all():
+                label = f"{content_type.app_label}.{content_type.model}"
+                labels.append(label)
+                template_content.append(tab_factory(label))
+            register_template_extensions(template_content)
+            self.features["template_extensions"] = sorted(set(labels))
+        except ProgrammingError:
+            logger.warning(
+                "Creating template content for validation engine failed because "
+                "the ContentType table was not available or populated. This is normal "
+                "during the execution of the migration command for the first time."
+            )
 
 
 config = NautobotDataValidationEngineConfig  # pylint:disable=invalid-name
